♻️ Use only the response payload of the lambda as output to make it even simpler to follow the data flow
When setting `outputPath` we filter the output of the lambda to a specific field. In this case its Payload (e.g the response body of the lambda). This way we strip all the other information we have returned from the lambda, such as `SdkHttpMetadata`, `StatusCode` etc..

Usually you're only interested in the response payload so we only use that as output.

Since we have also x-ray enabled, we have enough debug/tracing/observability information even for the fact that this is just a demo.